refactor: towards fixing `AffineRepr`
We implemented a bunch of generic operations on `AffineElement`
back when we added it such that we could implement the pair
of upstream traits for projective and affine curves, now
called `CurveGroup` and `AffineRepr`. This commit modifies
those operations such that the various type bounds on those
two traits are satisfied (they changed upstream with 0.4.x
Arkworks).