refactor: implement our own TE config
Previously defining the cofactor was done on `AffineCurve`: this
let us set it to 1 since decaf377 provides a prime-order group.
However, the cofactor is now moved to the curve configuration
via the trait `CurveConfig`.

This means we can no longer use the upstream `ark_ed_on_bls12_377`
curve configuration type: we need to define our own and implement
the required traits, and then use that to define `EdwardsAffine`,
`EdwardsProjective`, and `EdwardsVar` (R1CS) instead of the upstream
provided definitions of those types.