ci: dedicated profile for release + debug_assert
We want to use the `--release` behavior, but doing so directly would
disable evaluation of `debug_assert` lines, so we'll selectively
override that one option.

Updates the "check" and "test" jobs to use the `--profile ci`.
Also ensures that "check" covers all targets, including benchmarks,
and accordingly cleans up an unused import in same.